July 9 2024 guild meeting updates
