--- src/app/quiz/quiz-questions-service/quiz-questions.models.ts
@@ -1,28 +1,0 @@
-export interface Option {
-  text: string;
-  correct: boolean;
-}
-
-export interface QuizQuestion {
-  questionText: string;
-  options: Option[];
-  explanation: string;
-}
-
-export interface ApiQuizResults {
-  questions: QuizQuestion[];
- }
-// export interface Option {
-//   text: string;
-// }
-// export interface CorrectOption {
-//   text: string;
-//   correct: boolean;
-// }
-
-// export interface QuizQuestion {
-//   questionText: string;
-//   options: (Option | CorrectOption)[];    // Cool way to use OR [of a certain data type]
-//   explanation: string;
-// }
-
